pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "google-news-mcp"
version = "0.1.0"
description = "MCP server for Google News RSS feeds with advanced search and filtering"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "feedparser>=6.0.0",
    "googlenewsdecoder>=0.1.7",
]

[project.scripts]
google-news-mcp = "google_news_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/google_news_mcp"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
